requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
httpx==0.28.1
SQLAlchemy==2.0.48
alembic==1.16.5
boto3==1.42.67
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.2
PyJWT==2.10.1
pwdlib[argon2]==0.3.0
psycopg[binary]==3.2.12
python-multipart>=0.0.9
streamlit==1.55.0
